Special Actions and Hitbox Controlling
A controller for storing and initializing an array of special actions is possible now. Data for special actions have an animation, state, and group of hitboxes to use. UI for scrolling through your group of special actions is there as well.

Finalizing the HitboxController on the player that allows for instantiating a group of hitboxes from the list of special actions