Dash State; State and UI Fixes
A new DashState has been added, along with the inputs to set the state.

The lockout of states has been fixed. Replaying the same animation would not reset it, causing the animation event to exit the state to not be replayed.

The action UI has been redone to fix issues with the current action and one shown in the UI being different.