Separate Pickup Collider, Dynamic UI, etc
The sphere collider for pickup is now a separate object that is not on the player layer.

The canvas in the Player Bundle now scales with screen size, with the UI objects being resized individually as a result.

Basic attacks now don't move you as far when close to a locked-on target.

Action and Attack data now load from Resources, making the game able to be built.

You can turn the Spin action :)